fix: change user id table column name

+++ pages/api/cron.js
@@ -7,7 +7,7 @@ export default function handler(req, res) {
   const anonDeleteResult = supabase
     .from("your_table")
     .delete()
-    .eq("id_user", "anon")
+    .eq("user_id", "anon")
     .lt("created_at", new Date(Date.now() - 24 * 60 * 60 * 1000).toISOString());
 
   console.log('Deleted records for "anon" users:', anonDeleteResult);
@@ -15,7 +15,7 @@ export default function handler(req, res) {
   const identifierDeleteResult = supabase
     .from("your_table")
     .delete()
-    .not("id_user", "eq", "anon")
+    .not("user_id", "eq", "anon")
     .lt(
       "created_at",
       new Date(Date.now() - 30 * 24 * 60 * 60 * 1000).toISOString(),
